Update nf-mmiscapi-mmioopen.md
Removed the Note that the mmioOpen function is deprecated. The function is not deprecated (see Windows Kits 10.0.22621.0). It would not make sense for this function to be deprecated while all the other mmio* functions are not deprecated, as this is the only function that can provide an HMMIO file handle for those functions.